Update: move xhrSendWithIdleTimeout to Xhr
