+++ src/util/__tests__/uploads-test.js
@@ -1,11 +1,6 @@
 import { withData } from 'leche';
 
-import {
-    toISOStringNoMS,
-    getFileLastModifiedAsISONoMSIfPossible,
-    tryParseJson,
-    xhrSendWithIdleTimeout
-} from '../uploads';
+import { toISOStringNoMS, getFileLastModifiedAsISONoMSIfPossible, tryParseJson } from '../uploads';
 
 const sandbox = sinon.sandbox.create();
 
@@ -58,7 +53,7 @@ describe('util/uploads', () => {
                 it('should return the properly formatted date when possible and return null otherwise', () => {
                     expect(getFileLastModifiedAsISONoMSIfPossible(file)).to.equal(expectedResult);
                 });
-            },
+            }
         );
     });
 
@@ -78,71 +73,7 @@ describe('util/uploads', () => {
                 it('should return correct results', () => {
                     expect(tryParseJson(str)).to.deep.equal(expectedResult);
                 });
-            },
+            }
         );
-    });
-
-    describe('xhrSendWithIdleTimeout()', () => {
-        describe('xhrSendWithIdleTimeout()', () => {
-            let xhr;
-
-            beforeEach(() => {
-                xhr = sandbox.useFakeXMLHttpRequest();
-            });
-
-            afterEach(() => {
-                xhr.restore();
-            });
-
-            it('should call send() on underlying XHR', () => {
-                const request = new XMLHttpRequest();
-                const data = {};
-                sandbox.mock(request).expects('send').withArgs(data);
-                xhrSendWithIdleTimeout(request, data, 1000);
-            });
-
-            it('should call abort() and callback on underlying XHR after timeout', () => {
-                const request = new XMLHttpRequest();
-                const data = {};
-                let calls = 0;
-
-                function callback() {
-                    calls += 1;
-                }
-
-                request.open('GET', 'fake', true);
-                xhrSendWithIdleTimeout(request, data, 100, callback);
-
-                setTimeout(() => {
-                    sandbox.mock(request).expects('abort');
-                    assert.equal(calls, 1, 'callback not called exactly once');
-                }, 2000);
-            });
-
-            it('should call abort() if loaded has not changed', () => {
-                const request = new XMLHttpRequest();
-                const data = {};
-
-                request.open('GET', 'fake', true);
-                xhrSendWithIdleTimeout(request, data, 100);
-
-                setTimeout(() => {
-                    sandbox.mock(request).expects('abort');
-                    request.upload.eventListeners.progress[0]({ loaded: 0 });
-                }, 100);
-            });
-
-            it('should not call abort() if there has been progress', () => {
-                const request = new XMLHttpRequest();
-                const data = {};
-
-                request.open('GET', 'fake', true);
-                xhrSendWithIdleTimeout(request, data, 100);
-                setTimeout(() => {
-                    sandbox.mock(request).expects('abort').never();
-                    request.upload.eventListeners.progress[0]({ loaded: 1 });
-                }, 100);
-            });
-        });
     });
 });
