+++ src/util/uploads.js
@@ -67,73 +67,4 @@ function tryParseJson(maybeJson: string): ?Object {
     }
 }
 
-/**
- * Returns a handler for setInterval used in xhrSendWithIdleTimeout()
- * 
- * @param {number} lastProgress 
- * @param {number} timeoutMs 
- * @param {XMLHttpRequest} xhr 
- * @param {function} clear 
- * @param {?function} onTimeout
- * @return {function}
- */
-function getXhrIdleIntervalHandler(
-    lastProgress: number,
-    timeoutMs: number,
-    xhr: XMLHttpRequest,
-    clear: Function,
-    onTimeout?: Function,
-): Function {
-    return () => {
-        if (Date.now() - lastProgress <= timeoutMs) {
-            return;
-        }
-
-        xhr.abort();
-        clear();
-
-        if (onTimeout) {
-            onTimeout();
-        }
-    };
-}
-
-/**
- * Executes an upload via XMLHTTPRequest and aborts it if there is no progress event for at least timeoutMs.
- * 
- * @param {XMLHttpRequest} xhr
- * @param {object} data - Will be passed to xhr.send()
- * @param {number} timeoutMs - idle timeout, in milliseconds.
- * @param {function} onTimeout - callback invoked when request has timed out
- * @return {void}
- */
-function xhrSendWithIdleTimeout(xhr: XMLHttpRequest, data: Date, timeoutMs: number, onTimeout?: Function): void {
-    let interval;
-    let lastLoaded = 0;
-    let lastProgress = Date.now();
-
-    xhr.upload.addEventListener('progress', (event: ProgressEvent) => {
-        if (event.loaded > lastLoaded) {
-            lastLoaded = event.loaded;
-            lastProgress = Date.now();
-        }
-    });
-
-    function clear(): void {
-        if (!interval) {
-            return;
-        }
-
-        clearInterval(interval);
-        interval = null;
-    }
-
-    xhr.addEventListener('abort', clear);
-    xhr.addEventListener('load', clear);
-    xhr.addEventListener('error', clear);
-
-    interval = setInterval(getXhrIdleIntervalHandler(lastProgress, timeoutMs, xhr, clear, onTimeout), 1000);
-    xhr.send(data);
-}
-
-export { toISOStringNoMS, getFileLastModifiedAsISONoMSIfPossible, tryParseJson, xhrSendWithIdleTimeout };
+export { toISOStringNoMS, getFileLastModifiedAsISONoMSIfPossible, tryParseJson };
